[Keyless Common] Move test witnes_gen logic into circuit directory.
