+++ keyless-common/Cargo.toml
@@ -25,4 +25,3 @@ num-bigint = { workspace = true }
 num-traits = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
-tempfile = { workspace = true }
